refactor <TextSyntax>
Related to a module resolution issue when bundling.
For some reason the COLORS constant was not ready at build time.
