+++ code/system.ui/react.common/src/ui/Text.Syntax/Tokenizer.ts
