+++ code/system.ui/react.common/src/ui/Text.Syntax/types.mts
@@ -1,4 +1,4 @@
-import { type t } from '../common.t';
+import type { t } from '../common.t';
 
 type Color = string | number;
